+++ test/test_pipeline.py
@@ -39,22 +39,9 @@ FILENAMES = [
 ]
 
 
-@pytest.fixture(scope="class")
-def data(request):
-    global DATADIR
-    DATADIR = request.config.cache.makedir("download")
-    geneplexus.download.download_select_data(
-        DATADIR,
-        "All",
-        "BioGRID",
-        "Embedding",
-        ["GO", "DisGeNet"],
-    )
-
-
 def test_download_exist(caplog):
     geneplexus.download.download_select_data(
-        DATADIR,
+        pytest.datadir,
         "All",
         "BioGRID",
         "Embedding",
@@ -72,11 +59,11 @@ class TestGenePlexusPipeline(unittest.TestCase):
     @pytest.mark.order(0)
     def test_filenames(self):
         for filename in FILENAMES:
-            self.assertTrue(osp.isfile(osp.join(DATADIR, filename)))
+            self.assertTrue(osp.isfile(osp.join(pytest.datadir, filename)))
 
     @pytest.mark.order(1)
     def test_init_geneplexus(self):
-        self.gp.file_loc = DATADIR
+        self.gp.file_loc = pytest.datadir
         input_genes_path = osp.join(HOMEDIR, "example", "input_genes.txt")
         input_genes = geneplexus.util.read_gene_list(input_genes_path)
         self.gp.load_genes(input_genes)
